iOS: libuv's loop invoked at a higher frequency.
On iOS, the libuv's loop was called only once every 10ms, while on
Android is called every 1ms.
That was causing a significant performance difference between the
two platforms.
Calling the libuv's loop every 1ms, when there are operations still
pending, is not the optimal solution, since it still introduces an
unnecessary delay.
Removing the delay bumps up the performance significantly, but it may
also cause a high CPU usage in some cases when the calling code uses
long timeouts for operation that are not completing.
Therefore, for now, we have choosen to adopt an intermediate solution
that provides an acceptable trade-off between better performance and
low battery life impact.

Code cleanup:
 - removed unused global variable.
 - removed extra LoopOnce call that is not necessary.